+++ Humain_vs_Zombie/src/model/Item.java
@@ -5,10 +5,10 @@ public class Item {
 	private Integer id_item;
 	private Integer id_room;
 	private String name;
-	private Integer combat_points;
-	private Integer resistance;
+	private double combat_points;
+	private double resistance;
 
-	public Item(Integer id_item, Integer id_room, String name, Integer combat_points, Integer resistance) {
+	public Item(Integer id_item, Integer id_room, String name, double combat_points, double resistance) {
 		super();
 		this.id_item = id_item;
 		this.id_room = id_room;
@@ -41,19 +41,19 @@ public class Item {
 		this.name = name;
 	}
 
-	public Integer getCombat_points() {
+	public double getCombat_points() {
 		return combat_points;
 	}
 
-	public void setCombat_points(Integer combat_points) {
+	public void setCombat_points(double combat_points) {
 		this.combat_points = combat_points;
 	}
 
-	public Integer getResistance() {
+	public double getResistance() {
 		return resistance;
 	}
 
-	public void setResistance(Integer resistance) {
+	public void setResistance(double resistance) {
 		this.resistance = resistance;
 	}
